use pid_t instead of long for pid variables. #7099 (casper)
git-svn-id: https://origsvn.digium.com/svn/asterisk/trunk@25488 65c4cc65-6c06-0410-ace0-fbb531ad65f3

+++ logger.c
@@ -722,7 +722,7 @@ void ast_log(int level, const char *file, int line, const char *function, const 
 	if (logfiles.event_log && level == __LOG_EVENT) {
 		va_start(ap, fmt);
 
-		fprintf(eventlog, "%s asterisk[%d]: ", date, getpid());
+		fprintf(eventlog, "%s asterisk[%ld]: ", date, (long)getpid());
 		vfprintf(eventlog, fmt, ap);
 		fflush(eventlog);
